UPDATE: 로그아웃 기능 unit test 업데이트

+++ backend/src/main/java/kr/mj/gollaba/auth/service/AuthService.java
@@ -43,6 +43,10 @@ public class AuthService implements UserDetailsService {
     }
 
     public void logout(String refreshToken) {
+        if (!userTokenRepository.existsByRefreshToken(refreshToken)) {
+            throw new GollabaException(GollabaErrorCode.NOT_EXIST_REFRESH_TOKEN);
+        }
+
         userTokenRepository.deleteByRefreshToken(refreshToken);
     }
 
@@ -53,5 +57,6 @@ public class AuthService implements UserDetailsService {
 
         return new PrincipalDetails(user);
     }
+
 
 }
